refactor(config): extract stepper-config.ts (R6)
